Add WhiteNoise to serve static files
This commit adds the [WhiteNoise][wn] plugin to serve static files from
the app server. This removes a step of the deployment process. The
tradeoff is heavier load on the Python process that's running the
server, but we're unlikely to ever be in that situation, so I think that
the simplified deployment is very much the right choice for this
project.

wn: http://whitenoise.evans.io/en/stable/